Completed name change refactor
